add logout command
