+++ packages/cli/src/platform/login.ts
@@ -4,7 +4,7 @@ import http from 'http'
 import { red, underline } from 'kleur/colors'
 import open from 'open'
 
-import { writeAuthConfig } from '../utils/platformAuth'
+import { writeAuthConfig } from '../utils/platform'
 
 // const CONSOLE_URL = `https://console.prisma.io`
 const CONSOLE_URL = `http://localhost:8788`
@@ -15,7 +15,7 @@ export class Login implements Command {
   }
 
   public async parse(): Promise<string> {
-    logger.log('Authenticating to Cloud Platform via browser')
+    console.log('Authenticating to Cloud Platform via browser')
 
     const server = http.createServer()
     const { port } = await listen(server, 0, '127.0.0.1')
